Move panelling from ivc into blanket module
