+++ eudemo/eudemo/blanket/designer.py
@@ -33,7 +33,7 @@ from bluemira.geometry.constants import VERY_BIG
 from bluemira.geometry.face import BluemiraFace
 from bluemira.geometry.tools import boolean_cut, make_polygon
 from bluemira.geometry.wire import BluemiraWire
-from eudemo.ivc.panelling import PanellingDesigner
+from eudemo.blanket.panelling import PanellingDesigner
 from eudemo.tools import get_inner_cut_point
