+++ eudemo/eudemo_tests/blanket/panelling/data/panelling_ref_data.npy
